Fix same Enum.IsDefined type mismatch in BrokerRequest DTO/binding model
Same bug as CustomerAccountDTO: byte-backed Status property checked
against int-backed BrokerRequestStatus via the two-arg Enum.IsDefined
overload, which requires an exact type match and throws otherwise.

+++ Application.MainBoundedContext.DTO/AccountsModule/BrokerRequestBindingModel.cs
@@ -70,7 +70,7 @@ namespace Application.MainBoundedContext.DTO.AccountsModule
         {
             get
             {
-                return Enum.IsDefined(typeof(BrokerRequestStatus), Status) ? EnumHelper.GetDescription((BrokerRequestStatus)Status) : string.Empty;
+                return Enum.IsDefined(typeof(BrokerRequestStatus), (int)Status) ? EnumHelper.GetDescription((BrokerRequestStatus)Status) : string.Empty;
             }
         }
